Listening for session_type changes
Tony wants to send a message to any customer trying to book a colouring appointment.

This code simulates listening for a User making that selection, as the option has been removed from the UI till the message can be made available.

We'll replace the console.log with a UI message of some sort, and uncomment there selection of the default message